Big purge of dead code related to R asts
