+++ inst/editor/src/components/Inputs/SettingsFormBuilder/FormBuilder.stories.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 
 import omit from "just-omit";
+import type { ShinyUiNode } from "ui-node-definitions/src/ShinyUiNode";
 
 import { FormBuilder } from "./FormBuilder";
 import type { SettingsUpdateAction } from "./SettingsInput/SettingsInput";
@@ -35,6 +36,7 @@ export const AutoBuild = () => {
 
   return (
     <FormBuilder
+      app_tree={{} as ShinyUiNode}
       nodePath={[]}
       node={{
         id: "actionButton",
