+++ inst/editor/src/state/app_info.ts
@@ -2,7 +2,6 @@ import type { PayloadAction } from "@reduxjs/toolkit";
 import { createSlice } from "@reduxjs/toolkit";
 import type { MessageToClientByPath } from "communication-types";
 import type { AppInfo } from "communication-types/src/AppInfo";
-import type { RawRInfo } from "r-bindings";
 import { useSelector } from "react-redux";
 import type { ShinyUiNode } from "ui-node-definitions/src/ShinyUiNode";
 import type { PlaceNodeArguments } from "ui-node-definitions/src/TreeManipulation/placeNode";
@@ -59,14 +58,6 @@ export const mainStateSlice = createSlice({
     SET_FULL_STATE: (tree, action: PayloadAction<{ state: MainStateOption }>) =>
       action.payload.state,
 
-    SET_INFO_FROM_R: (tree, action: PayloadAction<RawRInfo>) => {
-      // const full_info = raw_R_info_to_app_info(action.payload);
-      throw new Error("Temporarily disabled");
-      // return {
-      //   mode: "MAIN",
-      //   ...full_info,
-      // };
-    },
     // This will initialize a state while also making sure the arguments match
     // what we expect in the app
     SET_APP_INFO: (state, action: PayloadAction<AppInfo>) => {
@@ -146,7 +137,6 @@ export const {
   PLACE_NODE,
   DELETE_NODE,
   SET_APP_INFO,
-  SET_INFO_FROM_R,
   SET_ERROR,
   SET_FULL_STATE,
   SHOW_TEMPLATE_CHOOSER,
